+++ enoslib/local.py
@@ -2,7 +2,12 @@ from .objects import Host
 
 
 class LocalHost(Host):
-    """Representation of your local machine."""
+    """Representation of a local machine.
+
+    Args:
+        alias: alias for a local machine
+            must be unique
+    """
 
     def __init__(self, alias: str = "localhost"):
         super().__init__(